Extract LogCensor to utils
