+++ Cargo.toml
@@ -13,7 +13,6 @@ crate-type = ["cdylib", "lib"]
 pretend_to_be_macos = ["telio-model/pretend_to_be_macos"]
 
 [dependencies]
-blake3 = "1.5.1"
 cfg-if = "1.0.0"
 ffi_helpers = "0.3.0"
 h2 = "=0.3.26" # RUSTSEC-2024-0332
